// src/infrastructure/database/enums/providers.enums.ts
export enum ProvidersEnums {
  POST_MODEL = 'POST_MODEL',
  BLOG_MODEL = 'BLOG_MODEL',
  USER_MODEL = 'USER_MODEL',
  COMMENT_MODEL = 'COMMENT_MODEL',
  LIKE_STATUS_COMMENTS_MODEL = 'LIKE_STATUS_COMMENTS_MODEL',
  LIKE_STATUS_POSTS_MODEL = 'LIKE_STATUS_POSTS_MODEL',
  LAST_10SEC_MODEL = 'LAST_10SEC_MODEL',
  CONFIRM_CODE_MODEL = 'CONFIRM_CODE_MODEL',
  BL_REFRESH_JWT_MODEL = 'BL_REFRESH_JWT_MODEL',
  DEVICES_MODEL = 'DEVICES_MODEL',
}